sentiment-analysis-pipeline/requirements.txt
apache-airflow==2.10.5
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.1
wordcloud==1.9.3
pyyaml==6.0.1
emoji==2.10.0
python-dateutil==2.8.2
